refactor: dto / domain 분리
