+++ src/main/java/gift/domain/member/MemberRepository.java
@@ -1,27 +1,26 @@
-package gift.web;
+package gift.domain.member;
 
-import gift.web.dto.Member;
+import gift.web.dto.MemberDto;
 import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public class MemberDAO {
+public class MemberRepository {
     private JdbcTemplate jdbcTemplate;
-    public MemberDAO(JdbcTemplate jdbcTemplate) {
+    public MemberRepository(JdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
     }
 
-    public void insertMember(Member member) {
+    public void insertMember(MemberDto memberDto) {
         var sql = "insert into members (email, password) values(?, ?)";
-        jdbcTemplate.update(sql, member.email(), member.password());
+        jdbcTemplate.update(sql, memberDto.email(), memberDto.password());
     }
 
-    public Member getMemberByEmail(String email) {
+    public MemberDto getMemberByEmail(String email) {
         var sql = "select * from members where email = ?";
         return jdbcTemplate.queryForObject(
             sql,
-            (rs, rowNum) -> new Member(
+            (rs, rowNum) -> new MemberDto(
                 rs.getString("email"),
                 rs.getString("password")
             ),
