--- src/main/java/gift/web/dto/Member.java
@@ -1,4 +1,0 @@
-package gift.web.dto;
-
-public record Member(String email,
-                     String password) { }
