+++ src/main/java/gift/web/dto/MemberDto.java
@@ -1,0 +1,4 @@
+package gift.web.dto;
+
+public record MemberDto(String email,
+                        String password) { }
